Interface for infectious disease modelling using brms.

Interface for infectious disease modelling using brms.

idbrm(data, formula, family, priors, custom_stancode, dry = FALSE, ...)

Arguments

data
A data frame as prepared for modelling using prepare with a class associated with the model prepared for.

formula
A formula as defined using id_formula or as supported by brms::brm.

family
A observation model family as defined in brms.

priors
A list of priors as defined using brms or id_priors. Defaults to the the id_priors defined for the model class being fit.

custom_stancode
A list of stanvars used to define custom stancode in brms. By default uses the code designed for the model class being fit (as specified using id_stancode).

dry
Logical, defaults to TRUE. For testing purposes should just the stan code be output with not fitting done.

...
Additional arguments to pass to brms::brm.

Author

Sam Abbott

Examples

# \donttest{
# define some example data
library(data.table)
dt <- data.table(
   region = "France", cases = seq(10, 500, by = 10),
   date = seq(as.Date("2020-10-01"), by = "days", length.out = 50)
   )
dt[, deaths := as.integer(shift(cases, 5) * 0.1)]

#>     region cases       date deaths
#>  1: France    10 2020-10-01     NA
#>  2: France    20 2020-10-02     NA
#>  3: France    30 2020-10-03     NA
#>  4: France    40 2020-10-04     NA
#>  5: France    50 2020-10-05     NA
#>  6: France    60 2020-10-06      1
#>  7: France    70 2020-10-07      2
#>  8: France    80 2020-10-08      3
#>  9: France    90 2020-10-09      4
#> 10: France   100 2020-10-10      5
#> 11: France   110 2020-10-11      6
#> 12: France   120 2020-10-12      7
#> 13: France   130 2020-10-13      8
#> 14: France   140 2020-10-14      9
#> 15: France   150 2020-10-15     10
#> 16: France   160 2020-10-16     11
#> 17: France   170 2020-10-17     12
#> 18: France   180 2020-10-18     13
#> 19: France   190 2020-10-19     14
#> 20: France   200 2020-10-20     15
#> 21: France   210 2020-10-21     16
#> 22: France   220 2020-10-22     17
#> 23: France   230 2020-10-23     18
#> 24: France   240 2020-10-24     19
#> 25: France   250 2020-10-25     20
#> 26: France   260 2020-10-26     21
#> 27: France   270 2020-10-27     22
#> 28: France   280 2020-10-28     23
#> 29: France   290 2020-10-29     24
#> 30: France   300 2020-10-30     25
#> 31: France   310 2020-10-31     26
#> 32: France   320 2020-11-01     27
#> 33: France   330 2020-11-02     28
#> 34: France   340 2020-11-03     29
#> 35: France   350 2020-11-04     30
#> 36: France   360 2020-11-05     31
#> 37: France   370 2020-11-06     32
#> 38: France   380 2020-11-07     33
#> 39: France   390 2020-11-08     34
#> 40: France   400 2020-11-09     35
#> 41: France   410 2020-11-10     36
#> 42: France   420 2020-11-11     37
#> 43: France   430 2020-11-12     38
#> 44: France   440 2020-11-13     39
#> 45: France   450 2020-11-14     40
#> 46: France   460 2020-11-15     41
#> 47: France   470 2020-11-16     42
#> 48: France   480 2020-11-17     43
#> 49: France   490 2020-11-18     44
#> 50: France   500 2020-11-19     45
#>     region cases       date deaths

dt[is.na(deaths), deaths := 0]

#>     region cases       date deaths
#>  1: France    10 2020-10-01      0
#>  2: France    20 2020-10-02      0
#>  3: France    30 2020-10-03      0
#>  4: France    40 2020-10-04      0
#>  5: France    50 2020-10-05      0
#>  6: France    60 2020-10-06      1
#>  7: France    70 2020-10-07      2
#>  8: France    80 2020-10-08      3
#>  9: France    90 2020-10-09      4
#> 10: France   100 2020-10-10      5
#> 11: France   110 2020-10-11      6
#> 12: France   120 2020-10-12      7
#> 13: France   130 2020-10-13      8
#> 14: France   140 2020-10-14      9
#> 15: France   150 2020-10-15     10
#> 16: France   160 2020-10-16     11
#> 17: France   170 2020-10-17     12
#> 18: France   180 2020-10-18     13
#> 19: France   190 2020-10-19     14
#> 20: France   200 2020-10-20     15
#> 21: France   210 2020-10-21     16
#> 22: France   220 2020-10-22     17
#> 23: France   230 2020-10-23     18
#> 24: France   240 2020-10-24     19
#> 25: France   250 2020-10-25     20
#> 26: France   260 2020-10-26     21
#> 27: France   270 2020-10-27     22
#> 28: France   280 2020-10-28     23
#> 29: France   290 2020-10-29     24
#> 30: France   300 2020-10-30     25
#> 31: France   310 2020-10-31     26
#> 32: France   320 2020-11-01     27
#> 33: France   330 2020-11-02     28
#> 34: France   340 2020-11-03     29
#> 35: France   350 2020-11-04     30
#> 36: France   360 2020-11-05     31
#> 37: France   370 2020-11-06     32
#> 38: France   380 2020-11-07     33
#> 39: France   390 2020-11-08     34
#> 40: France   400 2020-11-09     35
#> 41: France   410 2020-11-10     36
#> 42: France   420 2020-11-11     37
#> 43: France   430 2020-11-12     38
#> 44: France   440 2020-11-13     39
#> 45: France   450 2020-11-14     40
#> 46: France   460 2020-11-15     41
#> 47: France   470 2020-11-16     42
#> 48: France   480 2020-11-17     43
#> 49: France   490 2020-11-18     44
#> 50: France   500 2020-11-19     45
#>     region cases       date deaths

dt <- prepare(
  dt, model = "convolution", location = "region", 
  primary = "cases", secondary = "deaths",
  )

# fit the convolution model using a Poisson observation model
fit <- idbrm(data = dt, family = poisson(link = "identity"))

#> Warning: The global prior 'normal(0, 1)' of class 'b' will not be used in the model as all related coefficients have individual priors already. If you did not set those priors yourself, then maybe brms has assigned default priors. See ?set_prior and ?get_prior for more details.

#> Warning: The global prior 'normal(0, 0.1)' of class 'b' will not be used in the model as all related coefficients have individual priors already. If you did not set those priors yourself, then maybe brms has assigned default priors. See ?set_prior and ?get_prior for more details.

#> Warning: The global prior 'normal(0, 0.1)' of class 'b' will not be used in the model as all related coefficients have individual priors already. If you did not set those priors yourself, then maybe brms has assigned default priors. See ?set_prior and ?get_prior for more details.

#> Compiling Stan program...

#> Start sampling

#> 
#> SAMPLING FOR MODEL '88187c5aae95321a0a9a332a2550384d' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000896 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 8.96 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 4.71786 seconds (Warm-up)
#> Chain 1:                4.35793 seconds (Sampling)
#> Chain 1:                9.07578 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL '88187c5aae95321a0a9a332a2550384d' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000583 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 5.83 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 4.77815 seconds (Warm-up)
#> Chain 2:                4.18692 seconds (Sampling)
#> Chain 2:                8.96508 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL '88187c5aae95321a0a9a332a2550384d' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000578 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 5.78 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 4.92871 seconds (Warm-up)
#> Chain 3:                4.10008 seconds (Sampling)
#> Chain 3:                9.02879 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL '88187c5aae95321a0a9a332a2550384d' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000546 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 5.46 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 4.91812 seconds (Warm-up)
#> Chain 4:                4.16724 seconds (Sampling)
#> Chain 4:                9.08536 seconds (Total)
#> Chain 4: 

# }